Added PatchCamelyon dataset compatibility; Converting code to TF 1.15;

+++ EM Routing/metrics.py
@@ -24,7 +24,7 @@ def accuracy(logits, labels):
     logits = tf.identity(logits, name="logits")
     labels = tf.identity(labels, name="labels")
     batch_size = int(logits.get_shape()[0])
-    logits_idx = tf.to_int32(tf.argmax(logits, axis=1))
+    logits_idx = tf.cast(tf.argmax(logits, axis=1), dtype=tf.int32)
     logits_idx = tf.reshape(logits_idx, shape=(batch_size,))
     correct_preds = tf.equal(tf.to_int32(labels), logits_idx)
     accuracy = (tf.reduce_sum(tf.cast(correct_preds, tf.float32))
